Various small fixes
