+++ packages/frontend/src/components/icons/DownloadArrow.tsx
@@ -1,0 +1,19 @@
+import React, { SVGAttributes } from 'react'
+
+export function DownloadArrow(props: SVGAttributes<SVGElement>) {
+  return (
+    <svg
+      width="17"
+      height="17"
+      viewBox="0 0 17 17"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        d="M7.66665 0.166748C7.20665 0.166748 6.83332 0.540081 6.83332 1.00008V7.66675H3.91665C3.80615 7.66675 3.70017 7.71065 3.62203 7.78879C3.54389 7.86693 3.49999 7.97291 3.49999 8.08341C3.5 8.19391 3.54391 8.29988 3.62206 8.37801C3.63039 8.38652 3.63907 8.39466 3.6481 8.40242L7.90266 12.4145L7.93033 12.4405C8.08458 12.5856 8.28825 12.6664 8.49999 12.6667C8.71172 12.6664 8.9154 12.5856 9.06965 12.4405L9.08104 12.4307C9.08378 12.4275 9.08649 12.4243 9.08918 12.421L13.3421 8.41056C13.3505 8.4033 13.3587 8.39569 13.3665 8.38778L13.373 8.38289C13.3747 8.38128 13.3763 8.37965 13.3779 8.37801C13.4561 8.29988 13.5 8.19391 13.5 8.08341C13.5 7.97291 13.4561 7.86693 13.3779 7.78879C13.2998 7.71065 13.1938 7.66675 13.0833 7.66675H10.1667V1.00008C10.1667 0.540081 9.79332 0.166748 9.33332 0.166748H8.49999H7.66665ZM0.999987 15.1667C0.889561 15.1652 0.779925 15.1856 0.677451 15.2268C0.574978 15.2679 0.48171 15.3291 0.403069 15.4066C0.324427 15.4841 0.261981 15.5765 0.219358 15.6784C0.176734 15.7803 0.154785 15.8896 0.154785 16.0001C0.154785 16.1105 0.176734 16.2199 0.219358 16.3217C0.261981 16.4236 0.324427 16.516 0.403069 16.5935C0.48171 16.6711 0.574978 16.7322 0.677451 16.7734C0.779925 16.8146 0.889561 16.835 0.999987 16.8334H16C16.1104 16.835 16.22 16.8146 16.3225 16.7734C16.425 16.7322 16.5183 16.6711 16.5969 16.5935C16.6755 16.516 16.738 16.4236 16.7806 16.3217C16.8232 16.2199 16.8452 16.1105 16.8452 16.0001C16.8452 15.8896 16.8232 15.7803 16.7806 15.6784C16.738 15.5765 16.6755 15.4841 16.5969 15.4066C16.5183 15.3291 16.425 15.2679 16.3225 15.2268C16.22 15.1856 16.1104 15.1652 16 15.1667H0.999987Z"
+        fill="white"
+      />
+    </svg>
+  )
+}
