+++ packages/frontend/src/components/navbar/Navbar.stories.tsx
@@ -33,6 +33,7 @@ function Template(props: TemplateProps) {
       showActivity
       showHiring
       showHiringBadge
+      showMultisigReport
       selectedPage={props.selectedPage}
       socialLinks={{
         discordLink: '#',
