+++ packages/frontend/src/pages/multisig-report/view/MultisigReportPage.tsx
@@ -8,7 +8,7 @@ import {
   NavbarProps,
 } from '../../../components'
 import { PageContent } from '../../../components/PageContent'
-import { PolygonLogo } from '../../../components/PolygonLogo'
+import { PolygonLogo } from '../../../components/report/PolygonLogo'
 import { ReportBanner } from '../../../components/report/ReportBanner'
 import { ReportDownloadButton } from '../../../components/report/ReportDownloadButton'
